Add context methods to load and remove static buffers
This method allows for caching and retrieving buffers. The name parameter is
the (case sensitive!) string used for lookup, and is also the filename used by
standard I/O to load it. The name should be UTF-8 encoded.

Note that the returned buffer handles are not reference counted. That is, if
you retreive the same buffer handle twice, the first call to removeBuffer will
get rid of it.