Set the source orientation with the B-Format extension
Unlike the extension, setting the orientation changes both the B-Format
orientation *and* the source's direction. Setting the source direction only
changes the direction and leaves the B-Format orientation alone.

Ideally an app would only ever need to set the orientation as there's basically
no difference in behavior compared to direction. A zero-length direction is
functionally identical to having the min/max cone angles at 360 degrees (which
is the default).