Use pImpl for the DeviceManager
Unlike other pImpl objects, this can only be constructed from a get() call or
another DeviceManager object (which itself must be constructed from get() or
another object, etc). It cannot be cleared to null or compared, nor can its
handle be retrieved, as it can be assumed all DeviceManager objects refer to
the same implementation instance.