Add const char* overrides for some methods to avoid string copies
Although the StringView itself doesn't copy strings, it can't guarantee the
string it's providing a view to is null-terminated. OpenAL, being a C API,
requires null-terminated char arrays for strings. Consequently, the StringView
needs to be copied into a temporary String whose c_str() method can provide the
necessary C-style string. This extra copying is needlessly wasteful if the
input was initially a C-style string to begin with.