Add a method to get the clock time from the device
This is currently a psuedo-clock based on the standard steady_clock, but with
provisions to "pause" when there are no contexts or when pauseDSP() is called.
Eventually, it will utilize an OpenAL extension to get the actual audio
hardware clock (or at least base it on how fast the audio device consumes
samples).