Add a reinterpret_as method to ArrayView
This allows an ArrayView of one type to be reinterpreted as an ArrayView of
another type. The number of elements is adjusted given the size of the new
element type and the number of bytes in the original view (any "partial
element" at the end is excluded from the new view). An exception is thrown if
the view starts at an address that's not a multiple of the new type's
alignment.